// front/src/shared/api/orval/models/subscription.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * AI manager
 * хуйня
 * OpenAPI spec version: 1.0.0
 */
import type { User } from './user';
import type { SubscriptionCurrency } from './subscriptionCurrency';
import type { SubscriptionType } from './subscriptionType';
import type { SubscriptionMember } from './subscriptionMember';
import type { ChatMessage } from './chatMessage';
import type { SubscriptionGroup } from './subscriptionGroup';

export interface Subscription {
  /** Subscription ID */
  id: number;
  /** ID of the user who owns the subscription */
  userId: number;
  /** The user who owns the subscription */
  user: User;
  /** Name of the subscribed service */
  serviceName: string;
  /** Subscription price */
  price: number;
  /**
   * Currency of the subscription
   */
  currency: SubscriptionCurrency;
  /** Description of the subscription */
  description?: string;
  /** Subscription expiration date */
  expiresAt: string;
  /** Subscription start date */
  startsAt: string;
  /** Is the subscription active */
  isActive: boolean;
  /** Type of the subscription */
  type: SubscriptionType;
  /** Creation date */
  createdAt: string;
  /** Last update date */
  updatedAt: string;
  /** Members of the subscription */
  members: SubscriptionMember[];
  /** Chat messages related to the subscription */
  chatMessages: ChatMessage[];
  /**
   * Group associated with the subscription (if any)
   * @nullable
   */
  group: SubscriptionGroup;
}